requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1

# Document processing
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.8
pdfplumber==0.10.3

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tqdm==4.66.1
numpy==1.24.3
pandas==2.1.4

# API and interface
fastapi==0.108.0
uvicorn==0.25.0
gradio==4.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
